const.js moved to utils
